+++ ui/.env.development
@@ -5,13 +5,13 @@ VITE_USER_NODE_ENV=development
 VITE_PUBLIC_PATH=/koko
 
 # Socek 路径
-VITE_KOKO_WS_URL="http://localhost:5001"
+VITE_KOKO_WS_URL="ws://localhost:5050"
 
 # API 接口路径
-VITE_KOKO_API_URL="http://localhost:8080"
+VITE_KOKO_API_URL="http://localhost:5050"
 
 # 静态资路径
-VITE_KOKO_STATIC_URL="http://localhost:8080"
+VITE_KOKO_STATIC_URL="http://localhost:5050"
 
 # 打包压缩后是否删除源文件
 VITE_BUILD_COMPRESS_DELETE_ORIGIN_FILE=false
